requirements.txt
numpy>=1.20.0,<2.0.0
pandas>=1.3.0
matplotlib>=3.3.0
scikit-learn>=0.24.0
# Note: Requires valid Gurobi license
gurobipy>=9.5.0
torch>=1.9.0
python-Levenshtein>=0.12.0
chardet>=4.0.0
scipy>=1.7.0
POT>=0.8.0
pympler>=0.9

